pyproject.toml
[tool.poetry]
name = "caller-app"
version = "0.1.0"
description = ""
authors = ["Leo Kukharau <kuhorev.leonid@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
django = "^3.1.2"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"
django-environ = "^0.4.5"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.4.0"
autopep8 = "^1.5.4"
twilio = "^6.46.0"
django_otp = "^1.0.2"
qrcode = "^6.1"
django-phonenumber-field = "^5.0.0"
phonenumbers = "^8.12.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
